+++ core/lib/sykli/github/source.ex
@@ -5,17 +5,19 @@ defmodule Sykli.GitHub.Source do
 
   @impl true
   def acquire(context, token, opts \\ []) do
-    impl =
-      Keyword.get(opts, :impl, Application.get_env(:sykli, :github_source_impl, __MODULE__.Real))
-
-    impl.acquire(context, token, opts)
+    source_impl(opts).acquire(context, token, opts)
   end
 
   @impl true
   def cleanup(path, opts \\ []) do
-    impl =
-      Keyword.get(opts, :impl, Application.get_env(:sykli, :github_source_impl, __MODULE__.Real))
+    source_impl(opts).cleanup(path, opts)
+  end
 
-    impl.cleanup(path, opts)
+  defp source_impl(opts) do
+    Keyword.get(
+      opts,
+      :source_impl,
+      Keyword.get(opts, :impl, Application.get_env(:sykli, :github_source_impl, __MODULE__.Real))
+    )
   end
 end